Update beneficiary info
