+++ public/locales/bg/campaigns.json
@@ -97,8 +97,8 @@
     "status": "Статус:",
     "messages": "Кампанията подкрепиха:",
     "sort": {
-      "title" : "Сортирай по:",
-      "date" : "Дата",
+      "title": "Сортирай по:",
+      "date": "Дата",
       "amount": "Дарение",
       "search": "Търси...",
       "noResults": "Не са открити резултати"
@@ -136,6 +136,7 @@
     },
     "beneficiary": {
       "name": "Бенефициент",
+      "label": "Бенефициент:",
       "add": "Добави бенефициент",
       "create": "Създай бенефициент"
     },
@@ -149,6 +150,7 @@
       "add": "Добавете организатор",
       "create": "Създайте организатор"
     },
+    "image-of": "Изображение на",
     "podkrepi-bg-operator": "Подкрепи.бг Оператор:",
     "have-a-question": "Имате въпрос?",
     "btn-connect": "Пишете ни",
