+++ public/locales/en/campaigns.json
@@ -90,8 +90,8 @@
     "others": "others",
     "messages": "Messages:",
     "sort": {
-      "title" : "Sort by:",
-      "date" : "Date",
+      "title": "Sort by:",
+      "date": "Date",
       "amount": "Amount",
       "search": "Search...",
       "noResults": "No messages found"
@@ -131,6 +131,7 @@
     },
     "beneficiary": {
       "name": "Beneficiary",
+      "label": "Beneficiary:",
       "add": "Add beneficiary",
       "create": "Create beneficiary"
     },
@@ -144,6 +145,7 @@
       "add": "Add organizer",
       "create": "Create organizer"
     },
+    "image-of": "Image of",
     "podkrepi-bg-operator": "Podkrepi.bg Operator:",
     "have-a-question": "Have a question?",
     "btn-connect": "Write to us",
